Use custom program struct instead of lists
Because programs were just implemented directly as lists, it was not
possible to distinguish between a program and a regular list in the
contracts, so a program could be manipulated as a list and a list could
be applied like a program. Now programs are a separate type that cannot
be examined from inside of the language.