Add custom syntax for binding programs and values
The body of a module is processed by the new syntax module to split it
into module forms and expression forms, and expand special forms to
Racket syntax. List forms are transformed explicitly into program values
after being wrapped in a lambda to allow nested defines to work, which
also means that there is no need for a #%app binding. Other expressions
are retained as before and streamed out as values.

Programs can now be bound to names with the 'define' syntax. This syntax
must be followed by two programs: the first is a program whose only
contents must be identifiers (a program name and its parameters), and
the second is the program that will be the body of the defined program.

Values can now be bound to names with the 'bind' syntax. This syntax
must be followed by two programs: the first is a program whose only
contents must be identifiers, and the second is a program in which those
identifiers will be bound to values. When applied, the binding will
consume as many values as there are identifiers.

Values can now also be bound to names with the 'let' syntax. This is
effectively equivalent to Racket's let-values, though as above the
binding and body are the following program arguments. Unlike let-values,
execution of this form will not fail if the appropriate number of values
is not supplied by each expression in the binding.