Allow stream to be called recursively
Rather than printing values out when streaming, the stream function now
returns the values directly (which, when loading a module interactively,
produces the same effect). The difference is that now stream can be
called anywhere as a normal expression, and it will process the
available stack and return every value that results. Crucially, if a
procedure is on the stack and there are not enough values left to
satisfy the arity of the procedure, the result will be a procedure
returned after all of the preceding values, curried with the values that
were available and ready to accept more inputs. This replaces the error
case before that reported that there were not enough values: now not
having enough values just pauses computation and allows it to resume
once more values are available. As a result, stream can be called from
inside of other procedures, and the program will continue consuming
values once it (that is, the curried procedure that its evaluation got
stuck on) escapes onto the outer stack.