Add require and provide syntax forms
These two forms correspond exactly to their Racket counterparts, right
down to the exact same spec syntax. This addition means that it is now
possible to directly interact with a lot of the Racket ecosystem without
needing to define intermediate Racket functions that are then exposed
through the main module. This could also make it possible to cut most of
the default definitions and redefine them in Kihi modules directly.