Use quasisyntax and syntax-parse where appropriate
This change cuts down on unnecessary uses of with-syntax in order to get
computed syntax into larger expressions by unquoting the syntax directly
into quasisyntax. The use of syntax-parse instead of with-syntax when
testing that the syntax has the appropriate shape provides much better
error messages on bad syntax, including adding a context expression that
shows the whole piece of syntax that has failed to parse.

The syntax expansion now also considers that there may not be enough
forms remaining to satisfy the arity of the syntax, and raises an
appropriate syntax error detailing how many more forms were expected.