Implement nested delayed executions
There was a bug in the implementation of delaying the execution of a
procedure when there are not enough values on the stack to apply it,
which prevented multiple different procedures from being delayed in one
use of run. This meant that a procedure that needed to be delayed
because it was waiting on the results from another delayed procedure
would complain about the arity of the runtime implementation. This
change fixes the initial complaint by returning the correct number of
values from the collect procedure, and then completes the implementation
of nested delaying of procedures.