Rework module structure and implementation
The language runtime now works on lazy streams, transforming the tail of
the program every time an executable form appears. Procedures are not
executed arbitrarily, they must now be wrapped in a 'execute' struct. A
procedure that is curried because it could not consume enough values
from the stack also returns an execute form, which is also usable as a
procedure, to allow for simple interaction between Kihi and Racket.

Most of the prelude exports now have simple tests with a simple port of
rackunit, and all of the forms should be documented.