Treat local bindings differently from definitions
Local bindings are now internally wrapped in a 'no-execute' struct,
which is just a thunk around a value that the runtime can simply unwrap
instead of applying. Wrapping a bound name in parentheses treats it like
a definition instead of a local binding, acting like a pattern match on
the body of a procedure (though it will also currently work on
non-procedure values, with the same effect as though it were not wrapped
at all).

Program forms that contain only a single value are also wrapped in the
no-execute struct instead of creating a distinct program, which allows
direct reference to procedures that would otherwise execute if named.
This obviates the need for the 'arity' syntax, since the Racket function
procedure-reduce-arity can now be applied to a function. It has been
replaced by the Kihi function with-arity, which reduces the arity of a
given function and then applied it.