Only apply no-execute to referenced procedures
The no-execute form was being applied to any form in parentheses with
one sub-form inside, with no extra expansion applied to the sub-form.
This was problematic for expressions like ((x)), which was expanded into
the (no-execute (x)), where the (x) expression was intended to be a Kihi
expression but interpreted as a Racket expression. It was also a problem
for expressions that have the same meaning in Kihi and Racket, but whose
value was not a procedure, because the expression was not wrapped in a
procedure in the expanded form. The intent of the specialised expansion
was that the Kihi expression (+) is the same as the Racket expression +,
so the former should expand into the latter, but the Kihi expression
(nil) is not the same as the Racket expression nil.

Now the sub-form is always expanded first, which ensures that sub-forms
that are syntactically invalid are handled, so (define) remains a syntax
error instead of producing a problem after expansion. If the resulting
expansion includes a single identifier as its sub-form, then it expands
into an expression that checks if the value is a procedure, and only
wraps the result in no-execute if it is.