Ensure list/map works in the correct order
The old implementation of list/map used a left fold, so it put the list
back together in reverse order. A unit test with more than element has
been added to catch this case. The mapping function is also applied
without the list accumulator underneath it, which means mapping
functions can freely interact with the stack to the right of the
application of list/map without any other values intervening.